move seaborn/matplotlib drawing to new package structure
